test(scheduler): split periodic-tick test into tests/scheduler package
Move river_periodic_scheduler_tick_test.go (the 13s periodic-tick pole)
into a new tests/scheduler package with its own integration_testdb
TestMain bridge so it clones its own template DB and overlaps the river
pole as a concurrent process. Helpers route through testsupport.
